requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=1.0.0
Pillow>=8.0.0
numpy>=1.20.0
tqdm>=4.60.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
onnx>=1.12.0
onnxruntime>=1.12.0 
PyWavelets>=1.4.1